work around type weirdness
Prevents the client and admin SDK's `FirestoreDataConverter` and `WithFieldValue` interfering with each other which resulted in `FieldValue` types being added to one or the other

Now, the appropriate SDK `FirestoreDataConverter` is kept within the `/firebase` or `/firebase.server` modules and the converter class that will be defined is more 'internal use' by this lib.

Instead of creating an instance of a converter directly with `new PersonConverter(converter)` a factory function is provided. It has the slight added benefit of not needing the converter to be imported which can now be kept internal.